Doc(): add http exceptions to doc
