+++ docs/api/05-excessões.md
@@ -1,0 +1,3 @@
+:::app.exceptions
+:::app.exceptions.crud_error
+:::app.exceptions.http_error_schema
